Module I.Precision

Global precision for the functions I.pr and I.pp.

val set : int option -> unit

Set the number of decimals used by pr and pp. If None, use as many digits as needed to accurately print the interval.

val get : unit -> int option

Return the decimal precision used by pp, if any.