Module I.U

Module undoing the redeclaration of usual infix operators +, +., etc. in case it is needed locally, while this module is open.

Example: I.(x + sin(of_int U.(n + 1))).

Restore standard integer and floating point operators.

val (~-) : int -> int

val (~+) : int -> int

val (+) : int -> int -> int

val (-) : int -> int -> int

val (*) : int -> int -> int

val (/) : int -> int -> int

val (~-.) : float -> float

val (~+.) : float -> float

val (+.) : float -> float -> float

val (-.) : float -> float -> float

val (*.) : float -> float -> float

val (/.) : float -> float -> float

val (**) : float -> float -> float

val sqrt : float -> float

val (=) : 'a -> 'a -> bool

val (<>) : 'a -> 'a -> bool

val (<) : 'a -> 'a -> bool

val (>) : 'a -> 'a -> bool

val (<=) : 'a -> 'a -> bool

val (>=) : 'a -> 'a -> bool